Get payroll playing nicely with shared
