+++ api/src/lib/server/stripe.ts
@@ -1,6 +1,7 @@
 import Stripe from 'stripe';
 import { STRIPE_SECRET_KEY } from '$env/static/private';
+import { STRIPE_API_VERSION } from '@soli/shared/config';
 
 export const stripe = new Stripe(STRIPE_SECRET_KEY, {
-	apiVersion: '2026-05-27.dahlia'
+	apiVersion: STRIPE_API_VERSION
 });
