+++ api/src/routes/users/[slug]/collections/+server.ts
@@ -1,5 +1,5 @@
 import { handlePostgrestQuery, supabase } from '$lib/server/supabase';
-import { TABLES } from '@soli/shared/config/index';
+import { TABLES } from '@soli/shared/config';
 import type { CollectionHydrated } from '@soli/shared/types';
 import type { RequestHandler } from '@sveltejs/kit';
