+++ dashboard/src/routes/+layout.ts
@@ -1,7 +1,7 @@
 import { createBrowserClient, createServerClient, isBrowser } from '@supabase/ssr';
 import type { LayoutLoad } from './$types';
 import { PUBLIC_SUPABASE_ANON_KEY, PUBLIC_SUPABASE_URL } from '$lib/config';
-import { cookieOptions } from '@soli/shared/utils/cookies';
+import { cookieOptions } from '@soli/shared/config';
 
 export const load: LayoutLoad = async ({ fetch, data, depends }) => {
 	depends('supabase:auth');
