+++ shared/package.json
@@ -1,6 +1,14 @@
 {
 	"name": "@soli/shared",
 	"version": "1.0.0",
+	"exports": {
+		"./config": "./config/index.ts",
+		"./types": "./types/index.ts",
+		"./utils": "./utils/index.ts",
+		"./components": "./components/index.ts",
+		"./styles/global.css": "./styles/global.css",
+		"./styles/reset.css": "./styles/reset.css"
+	},
 	"devDependencies": {
 		"svelte": "^5.55.0"
 	}
